pyproject.toml
[project]
name = "stickerradar-classifier"
version = "0.1.0"
description = "Is this an image of a sticker?"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "clip",
    "fastapi>=0.115.12",
    "pillow>=11.2.1",
    "scikit-learn>=1.6.1",
    "torchvision>=0.22.0",
    "transformers>=4.51.3",
    "uvicorn>=0.34.2",
]

[tool.uv.sources]
clip = { git = "https://github.com/openai/CLIP.git" }
